3dunet/requirements.txt
torch
torchvision
numpy
scikit-learn
biopython
tensorboard
pyyaml
h5py